requirements.txt
requests
python-dotenv
dataclasses